% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fncut.R
\name{fncut}
\alias{fncut}
\title{fncut auxiliary converts continuous data into two categories}
\usage{
fncut(x)
}
\arguments{
\item{x}{{numerical vector of data values}}
}
\value{
x {vector of zeros and ones split at the median}
}
\description{
This is an internal function of the R package practicalSigni
 Psychologists use effect size to evaluate the practical
 importance of a treatment on a dependent variable using
 a binary [0,1] variable.  Assuming numerical data, we
 can always compute the median and regard values < or = the
 median as zero and other values as unity.
}
\author{
Prof. H. D. Vinod, Fordham University, NY
}
