% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapleyvalue.R
\name{shapleyvalue}
\alias{shapleyvalue}
\title{Shapley Value Regression importance of attributes in linear regression}
\usage{
shapleyvalue(y, x)
}
\arguments{
\item{y}{{input dependent variable data as a vector}}

\item{x}{{input matrix of p regressor variables must be a data frame}}
}
\value{
The structure of the output is a datatable, with two rows:
the unstandardized and standardized relative importance of each
attributes using shapley value regression method.
}
\description{
'ShapleyValue' package function shapleyvalue()
calculates the relative importance
of independent variables in linear regression while avoiding
collinearity. This function is a minor modification of the function
in the package ShapleyValue. The original function fails for me when
a variable name is x1. Since Jingyi Liang,
the maintainer of 'ShapleyValue,' has ignored several requests
to fix the error, this function
is reluctantly included here as a separate
function.
}
\references{
Liang J (2021). ShapleyValue: Shapley Value Regression for
Relative Importance of Attributes. R package version 0.2.0,
\url{https://CRAN.R-project.org/package=ShapleyValue}.
}
\author{
H. D. Vinod, Fordham University, NY
}
